Module hirofa_utils::js_utils

This contains abstract traits and structs for use with different javascript runtimes the Adapter traits are use in the worker thread (EventLoop) of the Runtime and thus are not Send, they should never leave the thread The facade classes are for use outside the worker thread, they are Send

Modules

adapters

facades

fetch
fetch api support

Structs

JsError

Script

Traits

ScriptPreProcessor