Function hirofa_utils::js_utils::adapters::promises::new_resolving_promise

pub fn new_resolving_promise<P, R, M, T>(
    realm: &T, 
    producer: P, 
    mapper: M
) -> Result<T::JsValueAdapterType, JsError> where
    T: JsRealmAdapter + 'static,
    R: Send + 'static,
    P: FnOnce() -> Result<R, JsError> + Send + 'static,
    M: FnOnce(&<<<<<<T as JsRealmAdapter>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeFacadeInnerType as JsRuntimeFacadeInner>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType, R) -> Result<<<<<<<<T as JsRealmAdapter>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeFacadeInnerType as JsRuntimeFacadeInner>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType as JsRealmAdapter>::JsValueAdapterType, JsError> + Send + 'static, 

create a new promise with a producer and a mapper the producer will run in a helper thread(in the tokio thread pool) and thus get a result asynchronously the resulting value will then be mapped to a JSValueRef by the mapper in the EventQueue thread the promise which was returned is then resolved with the value which is returned by the mapper