Trait hirofa_utils::js_utils::adapters::JsRealmAdapter

pub trait JsRealmAdapter {
    type JsRuntimeAdapterType: JsRuntimeAdapter;
    type JsValueAdapterType: JsValueAdapter + Clone + PartialEq;

    fn js_get_realm_id(&self) -> &str;

    fn js_get_runtime_facade_inner(
        &self
    ) -> Weak<<<<Self as JsRealmAdapter>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeFacadeInnerType>
    where
        Self: 'static;

    fn js_get_script_or_module_name(&self) -> Result<String, JsError>;

    fn js_eval(
        &self, 
        script: Script
    ) -> Result<Self::JsValueAdapterType, JsError>;

    fn js_proxy_install(
        &self, 
        proxy: JsProxy<Self>, 
        add_global_var: bool
    ) -> Result<Self::JsValueAdapterType, JsError>
    where
        Self: Sized;

    fn js_proxy_instantiate_with_id(
        &self, 
        namespace: &[&str], 
        class_name: &str, 
        instance_id: JsProxyInstanceId
    ) -> Result<Self::JsValueAdapterType, JsError>;

    fn js_proxy_instantiate(
        &self, 
        namespace: &[&str], 
        class_name: &str, 
        arguments: &[Self::JsValueAdapterType]
    ) -> Result<(JsProxyInstanceId, Self::JsValueAdapterType), JsError>;

    fn js_proxy_dispatch_event(
        &self, 
        namespace: &[&str], 
        class_name: &str, 
        proxy_instance_id: &JsProxyInstanceId, 
        event_id: &str, 
        event_obj: &Self::JsValueAdapterType
    ) -> Result<bool, JsError>;

    fn js_proxy_dispatch_static_event(
        &self, 
        namespace: &[&str], 
        class_name: &str, 
        event_id: &str, 
        event_obj: &Self::JsValueAdapterType
    ) -> Result<bool, JsError>;

    fn js_install_function(
        &self, 
        namespace: &[&str], 
        name: &str, 
        js_function: fn(_: &Self::JsRuntimeAdapterType, _: &Self, _: &Self::JsValueAdapterType, _: &[Self::JsValueAdapterType]) -> Result<Self::JsValueAdapterType, JsError>, 
        arg_count: u32
    ) -> Result<(), JsError>;

    fn js_install_closure<F: Fn(&Self::JsRuntimeAdapterType, &Self, &Self::JsValueAdapterType, &[Self::JsValueAdapterType]) -> Result<Self::JsValueAdapterType, JsError> + 'static>(
        &self, 
        namespace: &[&str], 
        name: &str, 
        js_function: F, 
        arg_count: u32
    ) -> Result<(), JsError>;

    fn js_eval_module(
        &self, 
        script: Script
    ) -> Result<Self::JsValueAdapterType, JsError>;

    fn js_get_global(&self) -> Result<Self::JsValueAdapterType, JsError>;

    fn js_get_namespace(
        &self, 
        namespace: &[&str]
    ) -> Result<Self::JsValueAdapterType, JsError>;

    fn js_function_invoke_by_name(
        &self, 
        namespace: &[&str], 
        method_name: &str, 
        args: &[Self::JsValueAdapterType]
    ) -> Result<Self::JsValueAdapterType, JsError>;

    fn js_function_invoke_member_by_name(
        &self, 
        this_obj: &Self::JsValueAdapterType, 
        method_name: &str, 
        args: &[Self::JsValueAdapterType]
    ) -> Result<Self::JsValueAdapterType, JsError>;

    fn js_function_invoke(
        &self, 
        this_obj: Option<&Self::JsValueAdapterType>, 
        function_obj: &Self::JsValueAdapterType, 
        args: &[&Self::JsValueAdapterType]
    ) -> Result<Self::JsValueAdapterType, JsError>;

    fn js_function_create<F: Fn(&Self, &Self::JsValueAdapterType, &[Self::JsValueAdapterType]) -> Result<Self::JsValueAdapterType, JsError> + 'static>(
        &self, 
        name: &str, 
        js_function: F, 
        arg_count: u32
    ) -> Result<Self::JsValueAdapterType, JsError>;

    fn js_error_create(
        &self, 
        name: &str, 
        message: &str, 
        stack: &str
    ) -> Result<Self::JsValueAdapterType, JsError>;

    fn js_object_delete_property(
        &self, 
        object: &Self::JsValueAdapterType, 
        property_name: &str
    ) -> Result<(), JsError>;

    fn js_object_set_property(
        &self, 
        object: &Self::JsValueAdapterType, 
        property_name: &str, 
        property: &Self::JsValueAdapterType
    ) -> Result<(), JsError>;

    fn js_object_get_property(
        &self, 
        object: &Self::JsValueAdapterType, 
        property_name: &str
    ) -> Result<Self::JsValueAdapterType, JsError>;

    fn js_object_create(&self) -> Result<Self::JsValueAdapterType, JsError>;

    fn js_object_construct(
        &self, 
        constructor: &Self::JsValueAdapterType, 
        args: &[&Self::JsValueAdapterType]
    ) -> Result<Self::JsValueAdapterType, JsError>;

    fn js_object_get_properties(
        &self, 
        object: &Self::JsValueAdapterType
    ) -> Result<Vec<String>, JsError>;

    fn js_object_traverse<F, R>(
        &self, 
        object: &Self::JsValueAdapterType, 
        visitor: F
    ) -> Result<Vec<R>, JsError>
    where
        F: Fn(&str, &Self::JsValueAdapterType) -> Result<R, JsError>;

    fn js_object_traverse_mut<F>(
        &self, 
        object: &Self::JsValueAdapterType, 
        visitor: F
    ) -> Result<(), JsError>
    where
        F: FnMut(&str, &Self::JsValueAdapterType) -> Result<(), JsError>;

    fn js_array_get_element(
        &self, 
        array: &Self::JsValueAdapterType, 
        index: u32
    ) -> Result<Self::JsValueAdapterType, JsError>;

    fn js_array_set_element(
        &self, 
        array: &Self::JsValueAdapterType, 
        index: u32, 
        element: &Self::JsValueAdapterType
    ) -> Result<(), JsError>;

    fn js_array_get_length(
        &self, 
        array: &Self::JsValueAdapterType
    ) -> Result<u32, JsError>;

    fn js_array_create(&self) -> Result<Self::JsValueAdapterType, JsError>;

    fn js_array_traverse<F, R>(
        &self, 
        array: &Self::JsValueAdapterType, 
        visitor: F
    ) -> Result<Vec<R>, JsError>
    where
        F: Fn(u32, &Self::JsValueAdapterType) -> Result<R, JsError>;

    fn js_array_traverse_mut<F>(
        &self, 
        array: &Self::JsValueAdapterType, 
        visitor: F
    ) -> Result<(), JsError>
    where
        F: FnMut(u32, &Self::JsValueAdapterType) -> Result<(), JsError>;

    fn js_null_create(&self) -> Result<Self::JsValueAdapterType, JsError>;

    fn js_undefined_create(&self) -> Result<Self::JsValueAdapterType, JsError>;

    fn js_i32_create(
        &self, 
        val: i32
    ) -> Result<Self::JsValueAdapterType, JsError>;

    fn js_string_create(
        &self, 
        val: &str
    ) -> Result<Self::JsValueAdapterType, JsError>;

    fn js_boolean_create(
        &self, 
        val: bool
    ) -> Result<Self::JsValueAdapterType, JsError>;

    fn js_f64_create(
        &self, 
        val: f64
    ) -> Result<Self::JsValueAdapterType, JsError>;

    fn js_promise_create(
        &self
    ) -> Result<Box<dyn JsPromiseAdapter<Self>>, JsError>;

    fn js_promise_add_reactions(
        &self, 
        promise: &Self::JsValueAdapterType, 
        then: Option<Self::JsValueAdapterType>, 
        catch: Option<Self::JsValueAdapterType>, 
        finally: Option<Self::JsValueAdapterType>
    ) -> Result<(), JsError>;

    fn js_promise_cache_add(
        &self, 
        promise_ref: Box<dyn JsPromiseAdapter<Self>>
    ) -> usize;

    fn js_promise_cache_consume(
        &self, 
        id: usize
    ) -> Box<dyn JsPromiseAdapter<Self>>;

    fn js_cache_add(&self, object: &Self::JsValueAdapterType) -> i32;

    fn js_cache_dispose(&self, id: i32);

    fn js_cache_with<C, R>(&self, id: i32, consumer: C) -> R
    where
        C: FnOnce(&Self::JsValueAdapterType) -> R;

    fn js_cache_consume(&self, id: i32) -> Self::JsValueAdapterType;

    fn js_instance_of(
        &self, 
        object: &Self::JsValueAdapterType, 
        constructor: &Self::JsValueAdapterType
    ) -> bool;

    fn js_json_stringify(
        &self, 
        object: &Self::JsValueAdapterType, 
        opt_space: Option<&str>
    ) -> Result<String, JsError>;

    fn js_json_parse(
        &self, 
        json_string: &str
    ) -> Result<Self::JsValueAdapterType, JsError>;

    fn js_typed_array_uint8_create(
        &self, 
        buffer: Vec<u8>
    ) -> Result<Self::JsValueAdapterType, JsError>;

    fn js_typed_array_uint8_create_copy(
        &self, 
        buffer: &[u8]
    ) -> Result<Self::JsValueAdapterType, JsError>;

    fn js_typed_array_detach_buffer(
        &self, 
        array: &Self::JsValueAdapterType
    ) -> Result<Vec<u8>, JsError>;

    fn js_typed_array_copy_buffer(
        &self, 
        array: &Self::JsValueAdapterType
    ) -> Result<Vec<u8>, JsError>;

    fn js_proxy_instance_get_info(
        &self, 
        obj: &Self::JsValueAdapterType
    ) -> Result<(String, JsProxyInstanceId), JsError>
    where
        Self: Sized;

    fn to_js_value_facade(
        &self, 
        js_value: &Self::JsValueAdapterType
    ) -> Result<JsValueFacade, JsError>
    where
        Self: Sized + 'static,
    { ... }

    fn from_js_value_facade(
        &self, 
        value_facade: JsValueFacade
    ) -> Result<Self::JsValueAdapterType, JsError>
    where
        Self: Sized + 'static,
    { ... }

    fn js_function_create_async<R, F>(
        &self, 
        name: &str, 
        js_function: F, 
        arg_count: u32
    ) -> Result<Self::JsValueAdapterType, JsError>
    where
        Self: Sized + 'static,
        R: Future<Output = Result<JsValueFacade, JsError>> + Send + 'static,
        F: Fn(JsValueFacade, Vec<JsValueFacade>) -> R + 'static,
    { ... }

    fn js_array_push_element(
        &self, 
        array: &Self::JsValueAdapterType, 
        element: &Self::JsValueAdapterType
    ) -> Result<u32, JsError> { ... }

    fn js_promise_create_resolving_async<P, R: Send + 'static, M>(
        &self, 
        producer: P, 
        mapper: M
    ) -> Result<Self::JsValueAdapterType, JsError>
    where
        P: Future<Output = Result<R, JsError>> + Send + 'static,
        M: FnOnce(&<<<<<<Self as JsRealmAdapter>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeFacadeInnerType as JsRuntimeFacadeInner>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType, R) -> Result<<<<<<<<Self as JsRealmAdapter>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeFacadeInnerType as JsRuntimeFacadeInner>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType as JsRealmAdapter>::JsValueAdapterType, JsError> + Send + 'static,
        Self: Sized + 'static,
    { ... }

    fn js_promise_create_resolving<P, R: Send + 'static, M>(
        &self, 
        producer: P, 
        mapper: M
    ) -> Result<Self::JsValueAdapterType, JsError>
    where
        P: FnOnce() -> Result<R, JsError> + Send + 'static,
        M: FnOnce(&<<<<<<Self as JsRealmAdapter>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeFacadeInnerType as JsRuntimeFacadeInner>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType, R) -> Result<<<<<<<<Self as JsRealmAdapter>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeFacadeInnerType as JsRuntimeFacadeInner>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType as JsRealmAdapter>::JsValueAdapterType, JsError> + Send + 'static,
        Self: Sized + 'static,
    { ... }

    fn js_instance_of_by_name(
        &self, 
        object: &Self::JsValueAdapterType, 
        constructor_name: &str
    ) -> Result<bool, JsError> { ... }

}

Associated Types

type JsRuntimeAdapterType: JsRuntimeAdapter

type JsValueAdapterType: JsValueAdapter + Clone + PartialEq

Required methods

fn js_get_realm_id(&self) -> &str

get the id of this Realm

fn js_get_runtime_facade_inner(
&self
) -> Weak<<<<Self as JsRealmAdapter>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeFacadeInnerType> where
Self: 'static,

get a Weak reference to the JsRuntimeFacadeInner this can be used to e.g. add a job to that JsRuntimeFacadeInner for resolving promises async

fn js_get_script_or_module_name(&self) -> Result<String, JsError>

get the name of the current module or script for modules loaded via the HttpModuleLoader these will have a https:// prefix modules loaded via the FileSystemModuleLoader will have a file:// prefix

fn js_eval(&self, script: Script) -> Result<Self::JsValueAdapterType, JsError>

eval a script please only use this for debugging/testing purposes although most JS engines will return values as if really calling a eval() method some may not (e.g. StarLight requires a return statement)

fn js_proxy_install(
&self,
proxy: JsProxy<Self>,
add_global_var: bool
) -> Result<Self::JsValueAdapterType, JsError> where
Self: Sized,

install a JsProxy into this Realm

fn js_proxy_instantiate_with_id(
&self,
namespace: &[&str],
class_name: &str,
instance_id: JsProxyInstanceId
) -> Result<Self::JsValueAdapterType, JsError>

instantiate a JsProxy instance value based on an instance_id this instance_id will be passed to the JsProxy’s member methods

fn js_proxy_instantiate(
&self,
namespace: &[&str],
class_name: &str,
arguments: &[Self::JsValueAdapterType]
) -> Result<(JsProxyInstanceId, Self::JsValueAdapterType), JsError>

instantiate a JsProxy instance value an instance_id will be generated and returned this instance_id will be passed to the JsProxy’s member methods

fn js_proxy_dispatch_event(
&self,
namespace: &[&str],
class_name: &str,
proxy_instance_id: &JsProxyInstanceId,
event_id: &str,
event_obj: &Self::JsValueAdapterType
) -> Result<bool, JsError>

dispatch an event to a JsProxy instance

fn js_proxy_dispatch_static_event(
&self,
namespace: &[&str],
class_name: &str,
event_id: &str,
event_obj: &Self::JsValueAdapterType
) -> Result<bool, JsError>

dispatch a static event to a JsProxy

fn js_install_function(
&self,
namespace: &[&str],
name: &str,
js_function: fn(_: &Self::JsRuntimeAdapterType, _: &Self, _: &Self::JsValueAdapterType, _: &[Self::JsValueAdapterType]) -> Result<Self::JsValueAdapterType, JsError>,
arg_count: u32
) -> Result<(), JsError>

install a function into this realm

fn js_install_closure<F: Fn(&Self::JsRuntimeAdapterType, &Self, &Self::JsValueAdapterType, &[Self::JsValueAdapterType]) -> Result<Self::JsValueAdapterType, JsError> + 'static>(
&self,
namespace: &[&str],
name: &str,
js_function: F,
arg_count: u32
) -> Result<(), JsError>

install a function into this realm based on a closure

fn js_eval_module(
&self,
script: Script
) -> Result<Self::JsValueAdapterType, JsError>

evaluate a module

fn js_get_global(&self) -> Result<Self::JsValueAdapterType, JsError>

get the global object

fn js_get_namespace(
&self,
namespace: &[&str]
) -> Result<Self::JsValueAdapterType, JsError>

get a namespace obj, when not present parts will be created

fn js_function_invoke_by_name(
&self,
namespace: &[&str],
method_name: &str,
args: &[Self::JsValueAdapterType]
) -> Result<Self::JsValueAdapterType, JsError>

invoke a function by name

fn js_function_invoke_member_by_name(
&self,
this_obj: &Self::JsValueAdapterType,
method_name: &str,
args: &[Self::JsValueAdapterType]
) -> Result<Self::JsValueAdapterType, JsError>

invoke a member function of an object by name

fn js_function_invoke(
&self,
this_obj: Option<&Self::JsValueAdapterType>,
function_obj: &Self::JsValueAdapterType,
args: &[&Self::JsValueAdapterType]
) -> Result<Self::JsValueAdapterType, JsError>

invoke a Funtion

fn js_function_create<F: Fn(&Self, &Self::JsValueAdapterType, &[Self::JsValueAdapterType]) -> Result<Self::JsValueAdapterType, JsError> + 'static>(
&self,
name: &str,
js_function: F,
arg_count: u32
) -> Result<Self::JsValueAdapterType, JsError>

create a new Function

fn js_error_create(
&self,
name: &str,
message: &str,
stack: &str
) -> Result<Self::JsValueAdapterType, JsError>

fn js_object_delete_property(
&self,
object: &Self::JsValueAdapterType,
property_name: &str
) -> Result<(), JsError>

delete a property of an Object

fn js_object_set_property(
&self,
object: &Self::JsValueAdapterType,
property_name: &str,
property: &Self::JsValueAdapterType
) -> Result<(), JsError>

set a property of an Object

fn js_object_get_property(
&self,
object: &Self::JsValueAdapterType,
property_name: &str
) -> Result<Self::JsValueAdapterType, JsError>

get a property of an Object

fn js_object_create(&self) -> Result<Self::JsValueAdapterType, JsError>

create a new Object

fn js_object_construct(
&self,
constructor: &Self::JsValueAdapterType,
args: &[&Self::JsValueAdapterType]
) -> Result<Self::JsValueAdapterType, JsError>

invoke a constructor Function to create new Object

fn js_object_get_properties(
&self,
object: &Self::JsValueAdapterType
) -> Result<Vec<String>, JsError>

get all property names of an Object

fn js_object_traverse<F, R>(
&self,
object: &Self::JsValueAdapterType,
visitor: F
) -> Result<Vec<R>, JsError> where
F: Fn(&str, &Self::JsValueAdapterType) -> Result<R, JsError>,

traverse all properties of an Object

fn js_object_traverse_mut<F>(
&self,
object: &Self::JsValueAdapterType,
visitor: F
) -> Result<(), JsError> where
F: FnMut(&str, &Self::JsValueAdapterType) -> Result<(), JsError>,

traverse all properties of an Object with a FnMut

fn js_array_get_element(
&self,
array: &Self::JsValueAdapterType,
index: u32
) -> Result<Self::JsValueAdapterType, JsError>

get an element of an Array

fn js_array_set_element(
&self,
array: &Self::JsValueAdapterType,
index: u32,
element: &Self::JsValueAdapterType
) -> Result<(), JsError>

set an element of an Array

fn js_array_get_length(
&self,
array: &Self::JsValueAdapterType
) -> Result<u32, JsError>

get the length of an Array

fn js_array_create(&self) -> Result<Self::JsValueAdapterType, JsError>

create a new Array

fn js_array_traverse<F, R>(
&self,
array: &Self::JsValueAdapterType,
visitor: F
) -> Result<Vec<R>, JsError> where
F: Fn(u32, &Self::JsValueAdapterType) -> Result<R, JsError>,

traverse all objects in an Array

fn js_array_traverse_mut<F>(
&self,
array: &Self::JsValueAdapterType,
visitor: F
) -> Result<(), JsError> where
F: FnMut(u32, &Self::JsValueAdapterType) -> Result<(), JsError>,

traverse all objects in an Array

fn js_null_create(&self) -> Result<Self::JsValueAdapterType, JsError>

create a null value

fn js_undefined_create(&self) -> Result<Self::JsValueAdapterType, JsError>

create an undefined value

fn js_i32_create(&self, val: i32) -> Result<Self::JsValueAdapterType, JsError>

create a Number value based on an i32

fn js_string_create(
&self,
val: &str
) -> Result<Self::JsValueAdapterType, JsError>

create a String value

fn js_boolean_create(
&self,
val: bool
) -> Result<Self::JsValueAdapterType, JsError>

create a Boolean value

fn js_f64_create(&self, val: f64) -> Result<Self::JsValueAdapterType, JsError>

create a Number value based on an f64

fn js_promise_create(&self) -> Result<Box<dyn JsPromiseAdapter<Self>>, JsError>

create a new Promise this returns JsPromiseAdapter which can be turned into a JsValueAdapter but can also be used to fulfill the promise

fn js_promise_add_reactions(
&self,
promise: &Self::JsValueAdapterType,
then: Option<Self::JsValueAdapterType>,
catch: Option<Self::JsValueAdapterType>,
finally: Option<Self::JsValueAdapterType>
) -> Result<(), JsError>

add reactions to an existing Promise object

fn js_promise_cache_add(
&self,
promise_ref: Box<dyn JsPromiseAdapter<Self>>
) -> usize

cache a JsPromiseAdapter so it can be accessed later based on an id, while cached the JsPromiseAdapter object will not be garbage collected

fn js_promise_cache_consume(&self, id: usize) -> Box<dyn JsPromiseAdapter<Self>>

Consume a JsPromiseAdapter (remove from cache)

fn js_cache_add(&self, object: &Self::JsValueAdapterType) -> i32

cache a Object so it can be accessed later based on an id, while cached the Object will not be garbage collected

fn js_cache_dispose(&self, id: i32)

remove an Object from the cache

fn js_cache_with<C, R>(&self, id: i32, consumer: C) -> R where
C: FnOnce(&Self::JsValueAdapterType) -> R,

use an Object in the cache

fn js_cache_consume(&self, id: i32) -> Self::JsValueAdapterType

get and remove an Object from the cache

fn js_instance_of(
&self,
object: &Self::JsValueAdapterType,
constructor: &Self::JsValueAdapterType
) -> bool

check if a JsValueAdapter is an instance of another JsValueAdapter

fn js_json_stringify(
&self,
object: &Self::JsValueAdapterType,
opt_space: Option<&str>
) -> Result<String, JsError>

turn any JsValueAdapter into a JSON string

fn js_json_parse(
&self,
json_string: &str
) -> Result<Self::JsValueAdapterType, JsError>

parse a JSON string into a JsValueAdapter

fn js_typed_array_uint8_create(
&self,
buffer: Vec<u8>
) -> Result<Self::JsValueAdapterType, JsError>

create a new typed array

fn js_typed_array_uint8_create_copy(
&self,
buffer: &[u8]
) -> Result<Self::JsValueAdapterType, JsError>

create a new typed array

fn js_typed_array_detach_buffer(
&self,
array: &Self::JsValueAdapterType
) -> Result<Vec<u8>, JsError>

fn js_typed_array_copy_buffer(
&self,
array: &Self::JsValueAdapterType
) -> Result<Vec<u8>, JsError>

fn js_proxy_instance_get_info(
&self,
obj: &Self::JsValueAdapterType
) -> Result<(String, JsProxyInstanceId), JsError> where
Self: Sized,

Provided methods

fn to_js_value_facade(
&self,
js_value: &Self::JsValueAdapterType
) -> Result<JsValueFacade, JsError> where
Self: Sized + 'static,

convert a JSValueAdapter to a Send able JSValueFacade you’ll need this to pass values out of the JSRuntimeAdapter’s worker thread the other way around you’ll need from_js_value_facade to move values into the worker thread

fn from_js_value_facade(
&self,
value_facade: JsValueFacade
) -> Result<Self::JsValueAdapterType, JsError> where
Self: Sized + 'static,

convert a JSValueFacade into a JSValueAdapter you need this to move values into the worker thread from a different thread (JSValueAdapter cannot leave the worker thread)

fn js_function_create_async<R, F>(
&self,
name: &str,
js_function: F,
arg_count: u32
) -> Result<Self::JsValueAdapterType, JsError> where
Self: Sized + 'static,
R: Future<Output = Result<JsValueFacade, JsError>> + Send + 'static,
F: Fn(JsValueFacade, Vec<JsValueFacade>) -> R + 'static,

fn js_array_push_element(
&self,
array: &Self::JsValueAdapterType,
element: &Self::JsValueAdapterType
) -> Result<u32, JsError>

push an element into an Array

fn js_promise_create_resolving_async<P, R: Send + 'static, M>(
&self,
producer: P,
mapper: M
) -> Result<Self::JsValueAdapterType, JsError> where
P: Future<Output = Result<R, JsError>> + Send + 'static,
M: FnOnce(&<<<<<<Self as JsRealmAdapter>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeFacadeInnerType as JsRuntimeFacadeInner>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType, R) -> Result<<<<<<<<Self as JsRealmAdapter>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeFacadeInnerType as JsRuntimeFacadeInner>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType as JsRealmAdapter>::JsValueAdapterType, JsError> + Send + 'static,
Self: Sized + 'static,

create a new Promise with a Future which will run async and then resolve or reject the promise the mapper is used to convert the result of the future into a JSValueAdapter

fn js_promise_create_resolving<P, R: Send + 'static, M>(
&self,
producer: P,
mapper: M
) -> Result<Self::JsValueAdapterType, JsError> where
P: FnOnce() -> Result<R, JsError> + Send + 'static,
M: FnOnce(&<<<<<<Self as JsRealmAdapter>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeFacadeInnerType as JsRuntimeFacadeInner>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType, R) -> Result<<<<<<<<Self as JsRealmAdapter>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeFacadeInnerType as JsRuntimeFacadeInner>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType as JsRealmAdapter>::JsValueAdapterType, JsError> + Send + 'static,
Self: Sized + 'static,

create a new Promise with a FnOnce producer which will run async and then resolve or reject the promise the mapper is used to convert the result of the future into a JSValueAdapter

fn js_instance_of_by_name(
&self,
object: &Self::JsValueAdapterType,
constructor_name: &str
) -> Result<bool, JsError>

check if a JsValueAdapter is an instance of a constructor by name

Example

if realm.js_instance_of_by_name(val, "Date")? {
   // it's a date
}