Trait hirofa_utils::js_utils::adapters::JsRuntimeAdapter

pub trait JsRuntimeAdapter {
    type JsRealmAdapterType: JsRealmAdapter + 'static;
    type JsRuntimeFacadeType: JsRuntimeFacade;
    fn js_load_module_script(
        &self, 
        ref_path: &str, 
        path: &str
    ) -> Option<Script>;

    fn js_create_realm(
        &self, 
        id: &str
    ) -> Result<&Self::JsRealmAdapterType, JsError>;

    fn js_remove_realm(&self, id: &str);

    fn js_get_realm(&self, id: &str) -> Option<&Self::JsRealmAdapterType>;

    fn js_get_main_realm(&self) -> &Self::JsRealmAdapterType;

    fn js_add_realm_init_hook<H>(&self, hook: H) -> Result<(), JsError>
    where
        H: Fn(&Self, &Self::JsRealmAdapterType) -> Result<(), JsError> + 'static;
}

Associated Types

type JsRealmAdapterType: JsRealmAdapter + 'static

type JsRuntimeFacadeType: JsRuntimeFacade

Required methods

fn js_load_module_script(&self, ref_path: &str, path: &str) -> Option<Script>

this method can be used to load the script code for a module (via any ScriptModuleLoader)

fn js_create_realm(
&self,
id: &str
) -> Result<&Self::JsRealmAdapterType, JsError>

create a new Realm

fn js_remove_realm(&self, id: &str)

drop a Realm, please note that the Realm might not really be dropped until all promises have fulfilled please note that this should not be able to remove the main realm

fn js_get_realm(&self, id: &str) -> Option<&Self::JsRealmAdapterType>

get a Realm, if the realm does not exists None will be returned

fn js_get_main_realm(&self) -> &Self::JsRealmAdapterType

get the main realm, this realm is always present and cannot be removed

fn js_add_realm_init_hook<H>(&self, hook: H) -> Result<(), JsError> where
H: Fn(&Self, &Self::JsRealmAdapterType) -> Result<(), JsError> + 'static,

add a hook to add custom code for when a Realm is initialized when adding a hook it will also be called for existing realms including the main realm