Trait hirofa_utils::js_utils::facades::JsRuntimeBuilder

pub trait JsRuntimeBuilder {
    type JsRuntimeFacadeType: JsRuntimeFacade;
    fn js_build(self) -> Self::JsRuntimeFacadeType;

    fn js_runtime_init_hook<H: FnOnce(&Self::JsRuntimeFacadeType) -> Result<(), JsError> + Send + 'static>(
        self, 
        hook: H
    ) -> Self;

    fn js_realm_adapter_init_hook<H: Fn(&<<Self as JsRuntimeBuilder>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType, &<<<Self as JsRuntimeBuilder>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType) -> Result<(), JsError> + Send + 'static>(
        self, 
        hook: H
    ) -> Self;

    fn js_runtime_adapter_init_hook<H: FnOnce(&<<Self as JsRuntimeBuilder>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType) -> Result<(), JsError> + Send + 'static>(
        self, 
        hook: H
    ) -> Self;

    fn js_script_pre_processor<S: ScriptPreProcessor + Send + 'static>(
        self, 
        preprocessor: S
    ) -> Self;

    fn js_script_module_loader<S: ScriptModuleLoader<<<<Self as JsRuntimeBuilder>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType> + Send + 'static>(
        self, 
        module_loader: S
    ) -> Self;

    fn js_compiled_module_loader<S: CompiledModuleLoader<<<<Self as JsRuntimeBuilder>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType> + Send + 'static>(
        self, 
        module_loader: S
    ) -> Self;

    fn js_native_module_loader<S: NativeModuleLoader<<<<Self as JsRuntimeBuilder>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType> + Send + 'static>(
        self, 
        module_loader: S
    ) -> Self
    where
        Self: Sized;
}

Associated Types

type JsRuntimeFacadeType: JsRuntimeFacade

Required methods

fn js_build(self) -> Self::JsRuntimeFacadeType

construct a JsRuntimeFacade based on this builders options

fn js_runtime_init_hook<H: FnOnce(&Self::JsRuntimeFacadeType) -> Result<(), JsError> + Send + 'static>(
self,
hook: H
) -> Self

add a runtime init hook, this closure will be invoked when a JsRuntimeFacade is constructed (e.g. builder.build is called)

fn js_realm_adapter_init_hook<H: Fn(&<<Self as JsRuntimeBuilder>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType, &<<<Self as JsRuntimeBuilder>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType) -> Result<(), JsError> + Send + 'static>(
self,
hook: H
) -> Self

add a realm adapter init hook, this will be called every time a realm is initialized

fn js_runtime_adapter_init_hook<H: FnOnce(&<<Self as JsRuntimeBuilder>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType) -> Result<(), JsError> + Send + 'static>(
self,
hook: H
) -> Self

fn js_script_pre_processor<S: ScriptPreProcessor + Send + 'static>(
self,
preprocessor: S
) -> Self

add a script preprocessor

fn js_script_module_loader<S: ScriptModuleLoader<<<<Self as JsRuntimeBuilder>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType> + Send + 'static>(
self,
module_loader: S
) -> Self

add a module loader

fn js_compiled_module_loader<S: CompiledModuleLoader<<<<Self as JsRuntimeBuilder>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType> + Send + 'static>(
self,
module_loader: S
) -> Self

add a compiled_module loader

fn js_native_module_loader<S: NativeModuleLoader<<<<Self as JsRuntimeBuilder>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType> + Send + 'static>(
self,
module_loader: S
) -> Self where
Self: Sized,

add a native module loader (e.g. with proxy classes instead of a script)