Trait hirofa_utils::js_utils::facades::JsRuntimeFacade

pub trait JsRuntimeFacade {
    type JsRuntimeAdapterType: JsRuntimeAdapter;

    fn js_realm_create(&mut self, name: &str) -> Result<(), JsError>;

    fn js_realm_destroy(&mut self, name: &str) -> Result<(), JsError>;

    fn js_realm_has(&mut self, name: &str) -> Result<bool, JsError>;

    fn js_loop_sync<R: Send + 'static, C: FnOnce(&Self::JsRuntimeAdapterType) -> R + Send + 'static>(
        &self, 
        consumer: C
    ) -> R;

    fn js_loop<R: Send + 'static, C: FnOnce(&Self::JsRuntimeAdapterType) -> R + Send + 'static>(
        &self, 
        consumer: C
    ) -> Pin<Box<dyn Future<Output = R> + Send>>;

    fn js_loop_void<C: FnOnce(&Self::JsRuntimeAdapterType) + Send + 'static>(
        &self, 
        consumer: C
    );

    fn js_eval(
        &self, 
        realm_name: Option<&str>, 
        script: Script
    ) -> Pin<Box<dyn Future<Output = Result<Box<dyn JsValueFacade>, JsError>>>>;

    fn js_function_invoke_sync(
        &self, 
        realm_name: Option<&str>, 
        namespace: &[&str], 
        method_name: &str, 
        args: Vec<Box<dyn JsValueFacade>>
    ) -> Result<Box<dyn JsValueFacade>, JsError>;

    fn js_function_invoke(
        &self, 
        realm_name: Option<&str>, 
        namespace: &[&str], 
        method_name: &str, 
        args: Vec<Box<dyn JsValueFacade>>
    ) -> Pin<Box<dyn Future<Output = Result<Box<dyn JsValueFacade>, JsError>>>>;

    fn js_function_invoke_void(
        &self, 
        realm_name: Option<&str>, 
        namespace: &[&str], 
        method_name: &str, 
        args: Vec<Box<dyn JsValueFacade>>
    );

    fn js_loop_realm_sync<R: Send + 'static, C: FnOnce(&Self::JsRuntimeAdapterType, &<Self::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType) -> R + Send + 'static>(
        &self, 
        realm_name: Option<&str>, 
        consumer: C
    ) -> R { ... }

    fn js_loop_realm<R: Send + 'static, C: FnOnce(&Self::JsRuntimeAdapterType, &<Self::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType) -> R + Send + 'static>(
        &self, 
        realm_name: Option<&str>, 
        consumer: C
    ) -> Pin<Box<dyn Future<Output = R>>> { ... }

    fn js_loop_realm_void<C: FnOnce(&Self::JsRuntimeAdapterType, &<Self::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType) + Send + 'static>(
        &self, 
        realm_name: Option<&str>, 
        consumer: C
    ) { ... }

}

The JsRuntime facade is the main entry point to the JavaScript engine, it is thread safe and handles the logic for transferring data from and to the JsRuntimeAdapter

Associated Types

type JsRuntimeAdapterType: JsRuntimeAdapter

Loading content...
Required methods

fn js_realm_create(&mut self, name: &str) -> Result<(), JsError>

fn js_realm_destroy(&mut self, name: &str) -> Result<(), JsError>

fn js_realm_has(&mut self, name: &str) -> Result<bool, JsError>

fn js_loop_sync<R: Send + 'static, C: FnOnce(&Self::JsRuntimeAdapterType) -> R + Send + 'static>(
&self,
consumer: C
) -> R

fn js_loop<R: Send + 'static, C: FnOnce(&Self::JsRuntimeAdapterType) -> R + Send + 'static>(
&self,
consumer: C
) -> Pin<Box<dyn Future<Output = R> + Send>>

fn js_loop_void<C: FnOnce(&Self::JsRuntimeAdapterType) + Send + 'static>(
&self,
consumer: C
)

fn js_eval(
&self,
realm_name: Option<&str>,
script: Script
) -> Pin<Box<dyn Future<Output = Result<Box<dyn JsValueFacade>, JsError>>>>

eval a script, please note that eval should not be used for production code, you should always use modules or functions and invoke them eval will always need to parse script and some engines like StarLight even require a different syntax (return(1); vs (1);) If None is passed as realm_name the default Realm wil be used

fn js_function_invoke_sync(
&self,
realm_name: Option<&str>,
namespace: &[&str],
method_name: &str,
args: Vec<Box<dyn JsValueFacade>>
) -> Result<Box<dyn JsValueFacade>, JsError>

Invoke a function and block until the function is done If None is passed as realm_name the default Realm wil be used

fn js_function_invoke(
&self,
realm_name: Option<&str>,
namespace: &[&str],
method_name: &str,
args: Vec<Box<dyn JsValueFacade>>
) -> Pin<Box<dyn Future<Output = Result<Box<dyn JsValueFacade>, JsError>>>>

Invoke a function this returns a Future which will fulfill when the function is done If None is passed as realm_name the default Realm wil be used

fn js_function_invoke_void(
&self,
realm_name: Option<&str>,
namespace: &[&str],
method_name: &str,
args: Vec<Box<dyn JsValueFacade>>
)

Invoke a function without waiting for a result If None is passed as realm_name the default Realm wil be used

Loading content...
Provided methods

fn js_loop_realm_sync<R: Send + 'static, C: FnOnce(&Self::JsRuntimeAdapterType, &<Self::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType) -> R + Send + 'static>(
&self,
realm_name: Option<&str>,
consumer: C
) -> R

fn js_loop_realm<R: Send + 'static, C: FnOnce(&Self::JsRuntimeAdapterType, &<Self::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType) -> R + Send + 'static>(
&self,
realm_name: Option<&str>,
consumer: C
) -> Pin<Box<dyn Future<Output = R>>>

fn js_loop_realm_void<C: FnOnce(&Self::JsRuntimeAdapterType, &<Self::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType) + Send + 'static>(
&self,
realm_name: Option<&str>,
consumer: C
)

Loading content...