Trait hirofa_utils::js_utils::facades::JsRuntimeFacadeInner

pub trait JsRuntimeFacadeInner {
    type JsRuntimeFacadeType: JsRuntimeFacade;
    fn js_exe_rt_task_in_event_loop<R: Send + 'static, J: FnOnce(&<<Self as JsRuntimeFacadeInner>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType) -> R + Send + 'static>(
        &self, 
        task: J
    ) -> R;

    fn js_add_rt_task_to_event_loop<R: Send + 'static, J: FnOnce(&<<Self as JsRuntimeFacadeInner>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType) -> R + Send + 'static>(
        &self, 
        task: J
    ) -> Pin<Box<dyn Future<Output = R>>>;

    fn js_add_rt_task_to_event_loop_void<J: FnOnce(&<<Self as JsRuntimeFacadeInner>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType) + Send + 'static>(
        &self, 
        task: J
    );
}

Associated Types

type JsRuntimeFacadeType: JsRuntimeFacade

Required methods

fn js_exe_rt_task_in_event_loop<R: Send + 'static, J: FnOnce(&<<Self as JsRuntimeFacadeInner>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType) -> R + Send + 'static>(
&self,
task: J
) -> R

run a closure in the EventLoop with a &JsRuntimeAdapter as only param and await the result blocking

fn js_add_rt_task_to_event_loop<R: Send + 'static, J: FnOnce(&<<Self as JsRuntimeFacadeInner>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType) -> R + Send + 'static>(
&self,
task: J
) -> Pin<Box<dyn Future<Output = R>>>

run a closure in the EventLoop with a &JsRuntimeAdapter as only param and return a future which will fulfill after the closure has been called

fn js_add_rt_task_to_event_loop_void<J: FnOnce(&<<Self as JsRuntimeFacadeInner>::JsRuntimeFacadeType as JsRuntimeFacade>::JsRuntimeAdapterType) + Send + 'static>(
&self,
task: J
)

run a closure in the EventLoop with a &JsRuntimeAdapter as only param