Enum hirofa_utils::js_utils::facades::values::JsValueFacade

pub enum JsValueFacade {

    I32 {
        val: i32,
    },
    F64 {
        val: f64,
    },
    String {
        val: String,
    },
    Boolean {
        val: bool,
    },
    JsObject {
        cached_object: CachedJsObjectRef,
    },
    JsPromise {
        cached_promise: CachedJsPromiseRef,
    },
    JsArray {
        cached_array: CachedJsArrayRef,
    },
    JsFunction {
        cached_function: CachedJsFunctionRef,
    },
    Object {
        val: HashMap<String, JsValueFacade>,
    },
    Array {
        val: Vec<JsValueFacade>,
    },
    Promise {
        producer: Mutex<Option<Pin<Box<dyn Future<Output = Result<JsValueFacade, JsError>> + Send + 'static>>>>,
    },
    Function {
        name: String,
        arg_count: u32,
        func: Arc<Box<dyn Fn(&[JsValueFacade]) -> Result<JsValueFacade, JsError> + Send + Sync>>,
    },
    JsError {
        val: JsError,
    },
    ProxyInstance {
        namespace: &'static [&'static str],
        class_name: &'static str,
        instance_id: JsProxyInstanceId,
    },
    Null,
    Undefined,

}

The JsValueFacade is a Send-able representation of a value in the Script engine

Variants

I32

Fields of I32

val: i32

F64

Fields of F64

val: f64

String

Fields of String

val: String

Boolean

Fields of Boolean

val: bool

JsObject

Fields of JsObject

cached_object: CachedJsObjectRef

JsPromise

Fields of JsPromise

cached_promise: CachedJsPromiseRef

JsArray

Fields of JsArray

cached_array: CachedJsArrayRef

JsFunction

Fields of JsFunction

cached_function: CachedJsFunctionRef

Object

Fields of Object

val: HashMap<String, JsValueFacade>

Array

Fields of Array

val: Vec<JsValueFacade>

Promise

Fields of Promise

producer: Mutex<Option<Pin<Box<dyn Future<Output = Result<JsValueFacade, JsError>> + Send + 'static>>>>

Function

Fields of Function

name: Stringarg_count: u32func: Arc<Box<dyn Fn(&[JsValueFacade]) -> Result<JsValueFacade, JsError> + Send + Sync>>

JsError

Fields of JsError

val: JsError

ProxyInstance

Fields of ProxyInstance

namespace: &'static [&'static str]class_name: &'static strinstance_id: JsProxyInstanceId

Null

Undefined

Implementations

impl JsValueFacade

pub fn new_i32(val: i32) -> Self

pub fn new_f64(val: f64) -> Self

pub fn new_bool(val: bool) -> Self

pub fn new_str(val: &str) -> Self

pub fn new_string(val: String) -> Self

pub fn new_callback<F: Fn(&[JsValueFacade]) -> Result<JsValueFacade, JsError> + Send + Sync + 'static>(
callback: F
) -> Self

pub fn new_function<F: Fn(&[JsValueFacade]) -> Result<JsValueFacade, JsError> + Send + Sync + 'static>(
name: &str,
function: F,
arg_count: u32
) -> Self

pub fn new_promise<T, R, P, M>(producer: P) -> Self where
T: JsRealmAdapter,
P: Future<Output = Result<JsValueFacade, JsError>> + Send + 'static,

create a new promise with a producer which will run async in a threadpool

pub fn is_i32(&self) -> bool

pub fn is_f64(&self) -> bool

pub fn is_bool(&self) -> bool

pub fn is_string(&self) -> bool

pub fn get_i32(&self) -> i32

pub fn get_f64(&self) -> f64

pub fn get_bool(&self) -> bool

pub fn get_str(&self) -> &str

pub fn is_null_or_undefined(&self) -> bool

pub fn get_value_type(&self) -> JsValueType

pub fn stringify(&self) -> String