Trait hirofa_utils::js_utils::facades::JsRuntimeFacade

pub trait JsRuntimeFacade {
    type JsRuntimeAdapterType: JsRuntimeAdapter;
    type JsRuntimeFacadeInnerType: JsRuntimeFacadeInner + Send + Sync;

    fn js_get_runtime_facade_inner(
        &self
    ) -> Weak<Self::JsRuntimeFacadeInnerType>;

    fn js_realm_create(&mut self, name: &str) -> Result<(), JsError>;

    fn js_realm_destroy(&mut self, name: &str) -> Result<(), JsError>;

    fn js_realm_has(&mut self, name: &str) -> Result<bool, JsError>;

    fn js_loop_sync<R: Send + 'static, C: FnOnce(&Self::JsRuntimeAdapterType) -> R + Send + 'static>(
        &self, 
        consumer: C
    ) -> R;

    fn js_loop_sync_mut<R: Send + 'static, C: FnOnce(&mut Self::JsRuntimeAdapterType) -> R + Send + 'static>(
        &self, 
        consumer: C
    ) -> R;

    fn js_loop<R: Send + 'static, C: FnOnce(&Self::JsRuntimeAdapterType) -> R + Send + 'static>(
        &self, 
        consumer: C
    ) -> Pin<Box<dyn Future<Output = R> + Send>>;

    fn js_loop_void<C: FnOnce(&Self::JsRuntimeAdapterType) + Send + 'static>(
        &self, 
        consumer: C
    );

    fn js_loop_realm_sync<R: Send + 'static, C: FnOnce(&Self::JsRuntimeAdapterType, &<<Self as JsRuntimeFacade>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType) -> R + Send + 'static>(
        &self, 
        realm_name: Option<&str>, 
        consumer: C
    ) -> R;

    fn js_loop_realm<R: Send + 'static, C: FnOnce(&Self::JsRuntimeAdapterType, &<<Self as JsRuntimeFacade>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType) -> R + Send + 'static>(
        &self, 
        realm_name: Option<&str>, 
        consumer: C
    ) -> Pin<Box<dyn Future<Output = R>>>;

    fn js_loop_realm_void<C: FnOnce(&Self::JsRuntimeAdapterType, &<<Self as JsRuntimeFacade>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType) + Send + 'static>(
        &self, 
        realm_name: Option<&str>, 
        consumer: C
    );

    fn js_eval(
        &self, 
        realm_name: Option<&str>, 
        script: Script
    ) -> Pin<Box<dyn Future<Output = Result<JsValueFacade, JsError>>>>;

    fn js_eval_module(
        &self, 
        realm_name: Option<&str>, 
        script: Script
    ) -> Pin<Box<dyn Future<Output = Result<(), JsError>>>>;

    fn js_function_invoke_sync(
        &self, 
        realm_name: Option<&str>, 
        namespace: &[&str], 
        method_name: &str, 
        args: Vec<JsValueFacade>
    ) -> Result<JsValueFacade, JsError>;

    fn js_function_invoke(
        &self, 
        realm_name: Option<&str>, 
        namespace: &[&str], 
        method_name: &str, 
        args: Vec<JsValueFacade>
    ) -> Pin<Box<dyn Future<Output = Result<JsValueFacade, JsError>>>>;

    fn js_function_invoke_void(
        &self, 
        realm_name: Option<&str>, 
        namespace: &[&str], 
        method_name: &str, 
        args: Vec<JsValueFacade>
    );

}

The JsRuntime facade is the main entry point to the JavaScript engine, it is thread safe and handles the logic for transferring data from and to the JsRuntimeAdapter

Associated Types

type JsRuntimeAdapterType: JsRuntimeAdapter

type JsRuntimeFacadeInnerType: JsRuntimeFacadeInner + Send + Sync

Required methods

fn js_get_runtime_facade_inner(&self) -> Weak<Self::JsRuntimeFacadeInnerType>

obtain a Weak reference to the JsRuntimeFacadeInner, this is often used to add jobs to the EventLoop from async tasks

fn js_realm_create(&mut self, name: &str) -> Result<(), JsError>

create a new JavaScript realm or context

fn js_realm_destroy(&mut self, name: &str) -> Result<(), JsError>

remove a JavaScript realm or context

fn js_realm_has(&mut self, name: &str) -> Result<bool, JsError>

check if a realm is present

fn js_loop_sync<R: Send + 'static, C: FnOnce(&Self::JsRuntimeAdapterType) -> R + Send + 'static>(
&self,
consumer: C
) -> R

util method to add a job to the EventLoop, usually this is passed to the JsRuntimeFacadeInner.js_loop_sync

fn js_loop_sync_mut<R: Send + 'static, C: FnOnce(&mut Self::JsRuntimeAdapterType) -> R + Send + 'static>(
&self,
consumer: C
) -> R

util method to add a job to the EventLoop, usually this is passed to the JsRuntimeFacadeInner.js_loop_sync

fn js_loop<R: Send + 'static, C: FnOnce(&Self::JsRuntimeAdapterType) -> R + Send + 'static>(
&self,
consumer: C
) -> Pin<Box<dyn Future<Output = R> + Send>>

util method to add a job to the EventLoop, usually this is passed to the JsRuntimeFacadeInner.js_loop_sync

fn js_loop_void<C: FnOnce(&Self::JsRuntimeAdapterType) + Send + 'static>(
&self,
consumer: C
)

util method to add a job to the EventLoop, usually this is passed to the JsRuntimeFacadeInner.js_loop_void

fn js_loop_realm_sync<R: Send + 'static, C: FnOnce(&Self::JsRuntimeAdapterType, &<<Self as JsRuntimeFacade>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType) -> R + Send + 'static>(
&self,
realm_name: Option<&str>,
consumer: C
) -> R

util method to add a job to the EventLoop, usually this is passed to the JsRuntimeFacadeInner.js_loop_realm_sync

fn js_loop_realm<R: Send + 'static, C: FnOnce(&Self::JsRuntimeAdapterType, &<<Self as JsRuntimeFacade>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType) -> R + Send + 'static>(
&self,
realm_name: Option<&str>,
consumer: C
) -> Pin<Box<dyn Future<Output = R>>>

util method to add a job to the EventLoop, usually this is passed to the JsRuntimeFacadeInner.js_loop_realm

fn js_loop_realm_void<C: FnOnce(&Self::JsRuntimeAdapterType, &<<Self as JsRuntimeFacade>::JsRuntimeAdapterType as JsRuntimeAdapter>::JsRealmAdapterType) + Send + 'static>(
&self,
realm_name: Option<&str>,
consumer: C
)

util method to add a job to the EventLoop, usually this is passed to the JsRuntimeFacadeInner.js_loop_realm_void

fn js_eval(
&self,
realm_name: Option<&str>,
script: Script
) -> Pin<Box<dyn Future<Output = Result<JsValueFacade, JsError>>>>

evaluate a script, please note that eval should not be used for production code, you should always use modules or functions and invoke them eval will always need to parse script and some engines like StarLight even require a different syntax (return(1); vs (1);) If None is passed as realm_name the main Realm wil be used

fn js_eval_module(
&self,
realm_name: Option<&str>,
script: Script
) -> Pin<Box<dyn Future<Output = Result<(), JsError>>>>

eval a script, please note that eval should not be used for production code, you should always use modules or functions and invoke them eval will always need to parse script and some engines like StarLight even require a different syntax (return(1); vs (1);) If None is passed as realm_name the main Realm wil be used

fn js_function_invoke_sync(
&self,
realm_name: Option<&str>,
namespace: &[&str],
method_name: &str,
args: Vec<JsValueFacade>
) -> Result<JsValueFacade, JsError>

Invoke a function and block until the function is done If None is passed as realm_name the main Realm wil be used

fn js_function_invoke(
&self,
realm_name: Option<&str>,
namespace: &[&str],
method_name: &str,
args: Vec<JsValueFacade>
) -> Pin<Box<dyn Future<Output = Result<JsValueFacade, JsError>>>>

Invoke a function this returns a Future which will fulfill when the function is done If None is passed as realm_name the main Realm wil be used

fn js_function_invoke_void(
&self,
realm_name: Option<&str>,
namespace: &[&str],
method_name: &str,
args: Vec<JsValueFacade>
)

Invoke a function without waiting for a result This method may be used instead of js_function_invoke when you don’t want to block_on the future or can’t .await it If None is passed as realm_name the main Realm wil be used